+++ GCECDev/Models/CarouselImageCaption.cs
@@ -1,0 +1,9 @@
+﻿using System;
+namespace GCECDev.Models
+{
+    public class CarouselImageCaption
+    {
+        public string ImageSrc { get; set; }
+        public string Caption { get; set; }
+    }
+}